+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVAWHG93T27S9NX4VN3GEM67"
+        "01KVAY6VC7XXYJTQS5KHYAXSB3"
     )
     r"""The unique identifier of the data source"""