+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVAWHG96GRA2GKGNZAJM4132"
+        "01KVAY6VCA90JENYJ2QG6GSPE5"
     )
     r"""The unique identifier of the data source"""